Pages Edited

+++ src/Pusula.Training.HealthCare.Domain.Shared/Salaries/SalaryConst.cs
@@ -15,12 +15,12 @@ namespace Pusula.Training.HealthCare.Salaries
             return string.Format(DefaultSorting, withEntityName ? "Salary." : string.Empty);
         }
 
-        public const decimal BaseAmountMinLength = 0;
-        public const decimal BaseAmountMaxLength = 1000000;
-        public const decimal BonusMinLength = 0;
-        public const decimal BonusMaxLength = 1000000;
-        public const decimal DeductionMinLength = 0;
-        public const decimal DeductionMaxLength = 1000000;
+        public const int BaseAmountMinLength = 0;
+        public const int BaseAmountMaxLength = 1000000;
+        public const int BonusMinLength = 0;
+        public const int BonusMaxLength = 1000000;
+        public const int DeductionMinLength = 0;
+        public const int DeductionMaxLength = 1000000;
 
     }
 }
